requirements.txt
pyirsdk==1.3.5
pandas==2.3.1
influxdb3-python==0.14.0